// teleaulaAppDev/src/app/interfaces/alumno.ts
export interface Alumno {
    id: number;
    nombre: string;
    apellidos: string;
    email: string;
    fechaNacimiento: string;
    localidad: string;
    provincia: string;
    codigoPostal: string;
    dni: string;
    numeroMatricula: string;
    fechaMatricula: string;
    ciclo: string;
    curso: string;
    direccion: string;
    asignaturasAlumno: [];
    alumnoEvaluacion: [];
 }